Tabs Props
Tab components now gets what to render based on the parent component it is called from.
