+++ client/src/components/Tabs.css
@@ -6,11 +6,12 @@
 .user-heading h2 {
   padding: 0.5rem 0rem;
   font-size: 20px;
+  margin-bottom: 2rem;
 }
 
 .tabs {
   display: flex;
-  margin-top: 2rem;
+  margin-top: 0rem;
   background: #36454f;
   box-shadow: rgba(60, 64, 67, 0.3) 0px 1px 2px 0px,
     rgba(60, 64, 67, 0.15) 0px 2px 6px 2px;
