+++ client/src/components/UserHeading.jsx
@@ -1,0 +1,11 @@
+import "./Tabs.css";
+
+const UserHeading = () => {
+  return (
+    <div className="user-heading">
+      <h2>Hello Ali,</h2>
+    </div>
+  );
+};
+
+export default UserHeading;
