--- client/src/pages/Generate.jsx
@@ -1,5 +1,0 @@
-const Generate = () => {
-  return <h2>Generate Timetable</h2>;
-};
-
-export default Generate;
