+++ client/src/pages/profile/About.jsx
@@ -1,9 +1,12 @@
 import Tabs from "../../components/Tabs";
+import UserHeading from "../../components/UserHeading";
+import { profileTabs } from "../TabsData";
 
 const About = () => {
   return (
     <>
-      <Tabs />
+      <UserHeading />
+      <Tabs tabs={profileTabs} />
       <h2>About Page</h2>
     </>
   );
